Fix partial routes causing `MissingRouteException`.
Moves converting URL arrays from the routing extension into the URL
voter, so that partial routes for the fuzzy route voter do not end up
being passed to `Router::url()`.

Except for the main URI that is ment to be used as the menu items
link, URLs are now being stored as is, and it's the voters
responsibility to handle them accordingly.

Refs #2